requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
attrs==25.3.0
backoff==2.2.1
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.39.4
botocore==1.39.4
branca==0.8.1
bs4==0.0.2
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
ExifRead==3.3.1
filelock==3.18.0
filetype==1.2.0
fire==0.7.0
folium==0.20.0
fonttools==4.58.5
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.73.1
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4
humanfriendly==10.0
idna==3.7
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kagglehub==0.3.12
kiwisolver==1.4.8
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.47.0
neo4j==5.28.1
networkx==3.4.2
numpy==2.2.6
onnx==1.18.0
onnx2torch==1.5.15
openai==1.88.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pgvector==0.4.1
pillow==11.2.1
pillow-avif-plugin==1.5.2
pillow_heif==1.0.0
plotly==6.1.2
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pybboxes==0.1.6
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
rawpy==0.25.0
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
reverse_geocoder==1.5.1
roboflow==1.2.1
rpds-py==0.26.0
s3transfer==0.13.0
sahi==0.11.30
scipy==1.15.3
seaborn==0.13.2
shapely==2.1.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
speciesnet==5.0.0
SQLAlchemy==2.0.41
st-pages==1.0.1
stack-data==0.6.3
streamlit==1.45.1
streamlit_folium==0.25.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.166
ultralytics-thop==2.0.14
umap-learn
urllib3==2.5.0
wcwidth==0.2.13
Werkzeug==3.1.3
xyzservices==2025.4.0
yolov5==7.0.11
langgraph
open-clip-torch
